Bug Fixing - Hazards bug -
When the Player jumps/falls on a hazard it loses two lives instead of one. Sometimes it even loses three.
This is something to do with the velocity of it hitting the hazard from above or it could be to do with the collision detection. It's a hell of a headache.

I posted a blog post it up and asked Nina what she could make of it and sent Johnny a link to this git hub repo. I'm going to go drink now. Fuck it all.