Improvements & Bug Fixing - Win Screen, Mouse Cursor,
I have built the simple Win Screen, improved the UI for the levels, main menu and win screen.
I have made it so that the mouse cursor only appears in the main menu.
The Enemy now just takes one life away.
The take life invoke was changed so that it's instant now.